chess_ai_management/requirements.txt
openai>=1.0.0
python-chess>=1.9.0
stockfish>=3.20.0
requests>=2.20.0  # For fetching FENs from GitHub if needed 